requirements_ale.txt
absl-py==2.1.0
acres==0.3.0
aiofiles==24.1.0
annotated-types==0.7.0
anthropic==0.18.1
anyio==3.7.1
astunparse==1.6.3
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.2.0
cryptography==44.0.2
cssselect==1.2.0
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.17.0
fitz==0.0.1.dev2
Flask==3.0.0
flatbuffers==25.2.10
frontend==0.0.3
fsspec==2025.3.0
gast==0.6.0
gevent==24.11.1
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-auth==2.38.0
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.69.1
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.62.3
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.2
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
keras==3.9.0
libclang==18.1.1
looseversion==1.3.0
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
narwhals==1.29.1
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==1.26.2
openai==1.65.4
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.5
Pillow==10.1.0
plotly==6.0.0
proto-plus==1.26.1
protobuf==4.25.6
prov==2.0.1
proxies==1.6
proxy==0.0.1
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
pydotplus==2.0.2
Pygments==2.19.1
pyparsing==3.2.1
pypdfium2==4.30.1
pyquery==2.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytils==0.4.1
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.31.0
rich==13.9.4
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.2
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
tools==0.1.9
tqdm==4.67.1
traits==7.0.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
zope.event==5.0
zope.interface==7.2
